Fix many warnings generated on Windows.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@774072 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/core/ActiveMQConsumer.h
@@ -331,7 +331,7 @@ namespace core{
          * @throws InvalidStateException if this consumer is closed upon
          * entering this method.
          */
-        Pointer<MessageDispatch> dequeue( int timeout )
+        Pointer<MessageDispatch> dequeue( long long timeout )
             throw ( cms::CMSException );
 
         /**
